// ANGULAR/ANGULAR/src/app/Shared/Model/Master/mla-constituency-master-model.ts
export class MLAConstituencyMasterModel {
  Id: number;
  Name: string;
  NameHindi: string;
  Latitude: string;
  Longitude: string;
  MAPImagePath: string;
  IsActive: boolean;
  DistrictCode: number|string;
  MLANameEng: string;
  MLANameHindi: string;
  MLAPhoto: string;
  ShortName: string;
  DesignationDisplayName: string;
  DesignationDisplayNameHindi: string;
  DesignationCode: number|string;
  DisplayName:string;
  URL:string;
  Email:string;
  whatsappNumber:string;
  FBLink:string;
  TwitterLink:string;
  DesignationShortName:string;
  DesignationShortNameHindi:string;
  MLAAddress: string;
}

export class MLAConstituencyViewModel{
  Id: number;
  Code: number;
  Name: string;
  NameHindi: string;
  Latitude: string;
  Longitude: string;
  MAPImagePath: string;
  IsActive: boolean;
  DistrictCode: number|string;
  MLANameEng: string;
  MLANameHindi: string;
  MLAPhoto: string;
  ShortName: string;
  DistrictTitle: string;
  ModifiedByName: string;
  ModifiedBy: Date|string;
  DesignationDisplayName: string;
  DesignationDisplayNameHindi: string;
  DesignationName: string;
  Email:string;
  whatsappNumber:string;
  FBLink:string;
  TwitterLink:string;
  DesignationShortName:string;
}